refactor(backend): extract the feature-status page into FeatureStatusReporter
Move the 370-line body of ConfigController::getFeaturesStatus() verbatim
into App\Service\Config\FeatureStatusReporter (S2/FM10, step 1). The
controller keeps the auth checks and returns the reporter's payload; the
two collaborators only that code used leave the controller constructor.

The only behavioural substitution is the HTTP probe: the private
file_get_contents helper becomes SidecarHealthProbeInterface, whose HTTP
implementation reproduces the same rules (2 s timeout, basic auth,
401/403 and 5xx down, 404 reachable) and is fakeable.

FeatureStatusReporterTest renders two fully faked environments (nothing
configured / everything healthy) and locks the payload byte-for-byte in
tests/Unit/Service/Config/__snapshots__. The snapshots were recorded on
this verbatim extraction so the next step — deriving rows from the
module registry — has a baseline to prove it changes nothing.
